hotfix - colon cancer loc_info augmentation
Former-commit-id: d671c6a5bb9d9a0297a6e932e18e686980fe2b9d

+++ datasets/colon_dataset.py
@@ -299,7 +299,7 @@ class ColonCancerBagsCross(data_utils.Dataset):
         for img in bag:
             if self.location_info:
                 bag_tensors.append(torch.cat(
-                    (img_transform(img[:, :, :3]),
+                    (img_transform(img[:, :, :3].astype('uint8')),
                      torch.from_numpy(img[:, :, 3:].astype(float).transpose((2, 0, 1))).float(),
                      )))
             else:
